+++ plugin-dev/Scripts/post-build-steps-mac.sh
@@ -11,7 +11,7 @@ ENGINE_DIR="$7"
 # Copy SentryObjC.dylib to plugin's Binaries/Mac dir if it doesn't exist there to ensure FAB version of the plugin works correctly
 # Epic obfuscates any extra binaries when pre-building the plugin
 if [ "$TARGET_PLATFORM" = "Mac" ] && [ ! -f "$PLUGIN_DIR/Binaries/Mac/SentryObjC.dylib" ]; then
-  cp "$PLUGIN_DIR/Source/ThirdParty/Mac/bin/SentryObjC.dylib" "$PLUGIN_DIR/Binaries/Mac/SentryObjC.dylib"
+  cp "$PLUGIN_DIR/Source/ThirdParty/Mac/Cocoa/bin/SentryObjC.dylib" "$PLUGIN_DIR/Binaries/Mac/SentryObjC.dylib"
 fi
 
 # Grant execute permissions to sentry-cli binary (FAB version of the plugin doesn't preserve file permissions)