+++ plugin-dev/Source/Sentry/Private/Interface/SentrySubsystemInterface.h
@@ -33,6 +33,7 @@ public:
 	virtual void InitWithSettings(const USentrySettings* settings, const FSentryCallbackHandlers& callbackHandlers) = 0;
 	virtual void Close() = 0;
 	virtual bool IsEnabled() = 0;
+	virtual bool IsCrashing() const = 0;
 	virtual ESentryCrashedLastRun IsCrashedLastRun() = 0;
 	virtual void AddBreadcrumb(TSharedPtr<ISentryBreadcrumb> breadcrumb) = 0;
 	virtual void AddBreadcrumbWithParams(const FString& Message, const FString& Category, const FString& Type, const TMap<FString, FSentryVariant>& Data, ESentryLevel Level) = 0;